Core Rewrite: fully-coupled solution Replaced the internal-ghost-node setting with a true sparse-matrix semi-implicit solution.
My error last time I tried this was in attempting to use the diagonal
elements of the adjacent stream segments directly as the off- (or on-)
diagonal inter-segment factors for the matrix X in the Ax = b solution.

This does not work, because this will link the head of one segment to the
head of another, due to their construction and indexing, alongside my use
of the roll() operator. (Indeed, this was an important clue before I went
through the math.)

Now, I calculate these boundaries separately, using the slope between the
final point for the upstream channel segment and the first point in the
downstream channel segment (both of which are, for convenience in generating
matrix calculations, though creating a bit of redundancy overall, stored
in the "extended" arrays).

The result of this rewrite is that I can take much longer time steps because
of the direct coupling across nodes -- and hence links -- in the network,
rather than needing to create a static ficticious internal-boundary ghost
node. This should help us to be able to perform more runs.