Build up `Network.initialize()` and supporting functions: LARGE EDIT. Set `initialize()` to require full set of objects to instantiate network.
This required new supporting functions as well, and changes to other functions (e.g., `LongProfile.set_x(...)`).

This is likely a first step towards changing the GRLP network interface, and may become a more generic interface than just for GRLP + tributary networks. We could use this approach to change river types (gravel, sand, ... lake (not a river)) in the midst of a network, and should be able to use it to set up different rules across the network.