Remaining problem in handling Q and slope. Iterations, too. Read on. In the test code, slopes smoothly vary if discharge smoothly varies.
Slopes also vary with discharge in a reasonable (directionally) seeming way, either within segments or directly at the junction.

Discharge variability is smoothed over 2 cells' width, with an intermediate cell "feeling" an intermediate discharge. *I wonder if this smoothing may have affected some of our tributary-junction handling, especially for the nodes just upstream of the junction, though this should be using just the upstream discharge now. This does create a sharp break.

Regardless of where the slope break is, the new slope is somewhat different from, and seemingly gentler than, the predicted slope. This matches the tributary-network case, in which the segment downstream of the confluence is simulated to be gentler than the rivers providing water to it, when it should otherwise maintain the same slope. The issue may be more the discharge handling and less the confluence handling.

Of note, the iterations now cause the result to vary somewhat wildly and then diverge. Perhaps the answer lies in this (1/6) power term, though I have not yet tested the question of not iterating.